+++ src/main/java/com/my/dto/TaskUpdateDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class TaskUpdateDto {
 
-    private Integer id; 
+    private Long id; 
     private String title;
     private String description;
     private Long assignedToId;
@@ -15,7 +15,7 @@ public class TaskUpdateDto {
     public TaskUpdateDto() {
     }
 
-    public TaskUpdateDto(Integer id, String title, String description, Long assignedToId, Long evaluatorId, LocalDate dueDate, String status) {
+    public TaskUpdateDto(Long id, String title, String description, Long assignedToId, Long evaluatorId, LocalDate dueDate, String status) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -25,11 +25,11 @@ public class TaskUpdateDto {
         this.status = status;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
